+++ src/redux/ReduxTypes.ts
@@ -1,17 +1,25 @@
-export type Thunk = (dispatch: Dispatch, getState: () => any) => void;
+import {Meta, MetaUpdate} from "../Types";
 
-export type Action<P = any> = {
+export type AsyncStateThunk = (dispatch: AsyncStateDispatch, getState: () => any) => void;
+
+export interface Action<P = any> {
   type: string;
   payload: P;
+}
+
+export type AsyncStateAction<P = any, E = void> = {
+  type: string;
+  payload: P;
+  extra: E | undefined;
 };
 
-export interface ActionCreator<P = any> {
-  (payload: P): Action<P>;
+export interface ActionCreator<P = any, E = void> {
+  (payload: P, extra?: E): AsyncStateAction<P, E>;
 
   type: string;
 }
 
-export interface AsyncActionCreators<P = any, V = any> {
+export interface AsyncActionCreators<P = any, V = any, M extends Meta = any> {
   (payload: P): void;
 
   reset: ActionCreator<void>;
@@ -20,34 +28,34 @@ export interface AsyncActionCreators<P = any, V = any> {
 
   refresh: ActionCreator<P>;
 
-  resolved: ActionCreator<V>;
+  resolved: ActionCreator<V, MetaUpdate<M>>;
 
   rejected: ActionCreator<Error>;
 
-  type: string;
-}
-
-export type Dispatch = (action: Action | Thunk) => Promise<any>;
-
-export interface AsyncActionWithThunk<P = any, V = any> {
-  (payload: P): Thunk;
+  pending: ActionCreator<void>;
 
   type: string;
 }
 
-export interface AsyncActionCreatorsWithThunk<P = any, V = any>
-  extends Omit<AsyncActionCreators<P, V>, "submit" | "refresh"> {
-  (payload: P): Thunk;
+export type AsyncStateDispatch = (action: Action | AsyncStateAction<any, any> | AsyncStateThunk) => Promise<any>;
 
-  submit: AsyncActionWithThunk<P, V>;
+export interface AsyncActionWithThunk<P = any, V = any, M extends Meta = any> {
+  (payload: P): AsyncStateThunk;
 
-  refresh: AsyncActionWithThunk<P, V>;
+  type: string;
+}
+
+export interface AsyncActionCreatorsWithThunk<P = any, V = any, M extends Meta = any>
+  extends Omit<AsyncActionCreators<P, V, M>, "submit" | "refresh"> {
+  (payload: P): AsyncStateThunk;
+
+  submit: AsyncActionWithThunk<P, V, M>;
+
+  refresh: AsyncActionWithThunk<P, V, M>;
 }
 
 export type AsyncActionHandler<S, P, V> = (
-  action: Action<P>,
-  dispatch: Dispatch,
+  action: AsyncStateAction<P, any>,
+  dispatch: AsyncStateDispatch,
   getState: () => S,
 ) => Promise<V>;
-
-export type TypedReducer<S> = (state: S, action: Action) => S;