+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 ## 0.9.0
 
+### Breaking
+
+- remove `StatefulPromise`
+- rename some redux types
+
+### Changes
+
+- add `meta` data to async states (eg. for when you want to store extra info about a state such as when it was cached)
 - `AsyncStateGroup` to allow efficient grouping of elements
 - improved error logging
-- add `pending` action to redux actions 
+- add `pending` action to `AsyncState` and redux actions 
 
 ## 0.8.0